Renamed folders to commit permission changes
